receipt/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
greenlet==3.0.1
asyncpg==0.30.0

# Computer Vision & OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0

# AI & ML
openai==1.3.7
langchain==0.0.350
sentence-transformers==2.2.2

# Configuration & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File Handling
aiofiles==23.2.0
python-multipart==0.0.6

# Date & Time
python-dateutil==2.8.2

# HTTP Client
httpx==0.25.2

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
structlog==23.2.0

# Environment
python-dotenv==1.0.0